large change on GUI - Rework GuiComponent, now it as Bacis Class for GUI - Rework Button static/Selector, Slider - added gui::CheckBox - now all states have IGUILayout field as basic GUI container for all GUI comnponent as Buttons, CheckBox, etc - was added FontManager for smart usage for states and  other components

+++ source/GUI/components/dropDownList.hpp
@@ -32,7 +32,7 @@ public:
 	// Functions
 	const bool getKeytime();
 	void updateKeytime(const float& dt);
-	void update(const sf::Vector2i& mousePosWindow, const float& dt);
+	void update(const sf::Vector2f& mousePosWindow, const float& dt);
 
 }; // namespace gui
 } // namespace gui
