+++ source/states/PauseMenu.hpp
@@ -1,8 +1,13 @@
 #ifndef CPP_STATE_PAUSE_HPP
 #define CPP_STATE_PAUSE_HPP
 
-#include "../GUI/components/button.hpp"
 #include "../core/header.h"
+
+#include "../core/tools/LOGGER.hpp"
+#include "../core/math/mymath.hpp"
+
+#include "../GUI/GUI_Component.hpp"
+#include "../GUI/components/button.hpp"
 
 class PauseMenu {
 private:
@@ -25,7 +30,7 @@ public:
 	// pos and size should be 0-100
 	void addButton(float btnPosY, sf::Vector2f btnSize, unsigned char_size, std::string key, std::string text);
 
-	void update(const sf::Vector2i& mousePosWindow);
+	void update(const sf::Vector2f& mousePosWindow);
 	void render(sf::RenderTarget& target);
 };
 #endif
